pyproject.toml
[project]
name = "pydantic-ai-for-fun"
version = "0.0.1"
description = "This is a template repository for Python projects that use uv for their dependency management."
authors = [{ name = "Jonathan Ndamba", email = "jonathan.ndamba.pro@gmaik.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "beautifulsoup4>=4.14.3",
    "duckduckgo-search>=8.1.1",
    "httpx>=0.28.1",
    "lancedb>=0.30.2",
    "logfire>=4.30.0",
    "loguru>=0.7.3",
    "mem0ai>=1.0.10",
    "pydantic-ai-slim[duckduckgo,web-fetch]>=1.73.0",
    "pydantic-ai>=1.73.0",
    "sentence-transformers>=5.3.0",
]

[project.urls]
Homepage = "https://jojo.github.io/pydantic-ai-for-fun/"
Repository = "https://github.com/jojo/pydantic-ai-for-fun"
Documentation = "https://jojo.github.io/pydantic-ai-for-fun/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.1",
    "tox-uv>=1.20.0",
    "deptry>=0.24.0",
    "ty>=0.0.23",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.5",
    "mkdocstrings[python]>=1.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pydantic_ai_for_fun"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["pydantic_ai_for_fun"]

